refactor: move components into own files
